refactor: extract MatchSnapshot to package-level function with runtime.Caller
